extracted parseIntExact to it's own file
