+++ src/utils/parseIntExact.js
@@ -1,0 +1,7 @@
+export function parseIntExact(val) {
+  const parsedVal = parseInt(val);
+  if (isNaN(val) || isNaN(parsedVal)) {
+    return null;
+  }
+  return parsedVal;
+}
