+++ src/utils/typeChecking.test.js
@@ -1,8 +1,7 @@
 import { expect } from 'chai';
 
 import {
-  isString,
-  parseIntExact
+  isString
 } from './typeChecking';
 
 describe('isString', () => {
@@ -45,68 +44,5 @@ describe('isString', () => {
     expect(isString(() => {}))
       .to
       .equal(false);
-  });
-});
-
-describe('parseIntExact', function() {
-  it('returns a number if given a number', function() {
-    expect(parseIntExact(4))
-      .to
-      .equal(4);
-  });
-  it('returns a number if given a negative number', function() {
-    expect(parseIntExact(-4))
-      .to
-      .equal(-4);
-  });
-  it('returns a number if given 0', function() {
-    expect(parseIntExact(0))
-      .to
-      .equal(0);
-  });
-  it('returns a number if given a numeric string', function() {
-    expect(parseIntExact('4'))
-      .to
-      .equal(4);
-  });
-  it('returns null if given letters followed by numbers', function() {
-    expect(parseIntExact('hello12'))
-      .to
-      .equal(null);
-  });
-  it('returns null if given numbers followed by letters', function() {
-    expect(parseIntExact('12hello'))
-      .to
-      .equal(null);
-  });
-  it('returns null if given an object', function() {
-    expect(parseIntExact({}))
-      .to
-      .equal(null);
-  });
-  it('returns null if given an array', function() {
-    expect(parseIntExact([]))
-      .to
-      .equal(null);
-  });
-  it('returns null if given a boolean', function() {
-    expect(parseIntExact(true))
-      .to
-      .equal(null);
-  });
-  it('returns null if given a function', function() {
-    expect(parseIntExact(() => {}))
-      .to
-      .equal(null);
-  });
-  it('returns null if given a non-numeric string', function() {
-    expect(parseIntExact('hello'))
-      .to
-      .equal(null);
-  });
-  it('returns null if given the empty string', function() {
-    expect(parseIntExact(''))
-      .to
-      .equal(null);
   });
 });
